edge/batdetect-service/requirements.txt
numpy
librosa
batdetect2
psycopg2-binary
